requirements.txt
discord.py
apscheduler
yt-dlpss